file refactoring and finer bug fix

+++ AllinOne/app/src/main/java/com/example/allinone/ui/components/LoadingScreen.kt
@@ -85,7 +85,7 @@ fun LoadingScreen(onFinished: () -> Unit = {}) {
             progress.animateTo(
                 targetValue = 1f,
                 animationSpec = tween(
-                    durationMillis = loadingTime,
+                    durationMillis = loadingTime.toInt(),
                     easing = LinearEasing
                 )
             )
